requirements.txt
streamlit
google-generativeai
edge-tts
requests
moviepy<2.0
python-dotenv
pypdf2
huggingface_hub
requests
edge-tts
